+++ src/backend/base/langflow/api/v2/workflow_reconstruction.py
@@ -10,9 +10,9 @@ from typing import TYPE_CHECKING
 
 from lfx.graph.graph.base import Graph
 from lfx.graph.schema import ResultData, RunOutputs
+from lfx.workflow.converters import run_response_to_workflow_response
 
 from langflow.api.v1.schemas import RunResponse
-from langflow.api.v2.converters import run_response_to_workflow_response
 from langflow.services.database.models.vertex_builds.crud import get_vertex_builds_by_job_id
 
 if TYPE_CHECKING:
