+++ src/backend/base/langflow/api/v2/workflow_validation.py
@@ -10,8 +10,8 @@ from __future__ import annotations
 
 from fastapi import HTTPException, status
 from lfx.utils.flow_validation import CustomComponentValidationError, validate_flow_for_current_settings
+from lfx.workflow.converters import ParsedWorkflowRun
 
-from langflow.api.v2.converters import ParsedWorkflowRun
 from langflow.services.authorization.fetch import deny_to_404
 from langflow.services.database.models.flow.model import FlowRead
 from langflow.services.database.models.user.model import UserRead
